Add VSDashboardGate and remove AuthCreateAccount
