--- frontend/src/pages/AuthCreateAccount/index.ts
@@ -1,1 +1,0 @@
-export * from './AuthCreateAccount';
